+++ src/System.Collections.Concurrent/src/Configurations.props
@@ -4,6 +4,8 @@
     <BuildConfigurations>
       netcoreapp;
       uap-Windows_NT;
+      netcoreapp-Windows_NT;
+      netcoreapp-Unix;
     </BuildConfigurations>
   </PropertyGroup>
 </Project>
